pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-redteam-eval"
version = "0.1.0"
description = "Systematic adversarial evaluation framework for LLMs and agents"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
dev = ["pytest>=7.0", "ruff>=0.8"]

[project.scripts]
redteam = "redteam.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["redteam", "adapters"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
